Update gravityDB Check
Per discussion in pi-hole/docker-pi-hole#883, sourcing
pihole-FTL.conf is bad practice and potentially dangerous.

Grepping it is the prefered solution. A line of verbosity is added
as a breadcrumb to revisit if/when pihole-FTL.conf is succeeded by
TOML. This is to safeguard against a silent regression of #3948
if/when pihole-FTL.conf is no longer used.

Signed-off-by: D.Rect <48034372+DistractionRectangle@users.noreply.github.com>